Updated commands to modern eco
