+++ eco-core/core-plugin/src/main/java/com/willfp/stattrackers/commands/CommandReload.java
@@ -6,7 +6,10 @@ import com.willfp.eco.core.command.impl.PluginCommand;
 import com.willfp.eco.core.command.impl.Subcommand;
 import com.willfp.eco.util.NumberUtils;
 import com.willfp.eco.util.StringUtils;
+import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
 
 public class CommandReload extends Subcommand {
     /**
@@ -19,12 +22,10 @@ public class CommandReload extends Subcommand {
     }
 
     @Override
-    public CommandHandler getHandler() {
-        return (sender, args) -> {
-            sender.sendMessage(
-                    this.getPlugin().getLangYml().getMessage("reloaded", StringUtils.FormatOption.WITHOUT_PLACEHOLDERS)
-                            .replace("%time%", NumberUtils.format(this.getPlugin().reloadWithTime()))
-            );
-        };
+    public void onExecute(@NotNull CommandSender sender, @NotNull List<String> args) {
+        sender.sendMessage(
+                this.getPlugin().getLangYml().getMessage("reloaded", StringUtils.FormatOption.WITHOUT_PLACEHOLDERS)
+                        .replace("%time%", NumberUtils.format(this.getPlugin().reloadWithTime()))
+        );
     }
 }
