+++ eco-core/core-plugin/src/main/java/com/willfp/stattrackers/commands/CommandStatTrackers.java
@@ -3,7 +3,10 @@ package com.willfp.stattrackers.commands;
 import com.willfp.eco.core.EcoPlugin;
 import com.willfp.eco.core.command.CommandHandler;
 import com.willfp.eco.core.command.impl.PluginCommand;
+import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
 
 public class CommandStatTrackers extends PluginCommand {
     /**
@@ -20,9 +23,7 @@ public class CommandStatTrackers extends PluginCommand {
     }
 
     @Override
-    public CommandHandler getHandler() {
-        return (sender, args) -> {
-            sender.sendMessage(this.getPlugin().getLangYml().getMessage("invalid-command"));
-        };
+    public void onExecute(@NotNull CommandSender sender, @NotNull List<String> args) {
+        sender.sendMessage(this.getPlugin().getLangYml().getMessage("invalid-command"));
     }
 }
